requirements.txt
certifi==2022.12.7
chardet==3.0.4
idna==2.8
jpholiday==0.1.8
pybacklogpy==0.12
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.22.0
six==1.16.0
urllib3==1.25.11
